+++ src/hdr_image.rs
@@ -12,14 +12,14 @@
 //!
 //! - Linear RGB floating-point storage
 //! - Safe pixel indexing with bounds checking
-//! - Basic tone mapping (Reinhard operator)
+//! - Basic tone mappings
 //! - Log-average luminance computation
 //!
 //! ## Example
 //!
 //! ```rust, no_run
-//! use crate::rstrace::color::Color;
-//! use crate::rstrace::hdr_image::HDR;
+//! use rstrace::color::Color;
+//! use rstrace::hdr_image::HDR;
 //!
 //! let mut img = HDR::new(512, 512);
 //!
@@ -31,14 +31,13 @@
 //!
 
 use crate::color::Color;
-use crate::functions::endianness_number;
+use crate::functions::{are_close, endianness_number};
 use anyhow::{Result, anyhow};
 use byteorder::{BigEndian, LittleEndian, WriteBytesExt};
-use endianness::ByteOrder;
 use std::fs::File;
 use std::io::{BufReader, Write};
 
-use crate::pfm_func::{Parameter, read_pfm};
+use crate::pfm_func::{Endianness, Parameter, read_pfm};
 use image::{Rgb, RgbImage};
 
 /// Represents an HDR (High Dynamic Range) image.
@@ -85,89 +84,6 @@ impl HDR {
         }
     }
 
-    /// Writes the image to a `.pfm` (Portable Float Map) file.
-    ///
-    /// # Arguments
-    /// * `writer` - The output destination (e.g., a file, a memory buffer, or a network socket).
-    ///   It accepts any type that implements the [`Write`] trait.
-    /// * `endianness` - Byte order used for writing floats
-    ///
-    /// # Errors
-    /// Returns an error if:
-    /// - The underlying stream (`writer`) cannot be written to.
-    /// - An I/O error occurs while formatting the header or encoding the pixels.
-    ///
-    /// # Examples
-    ///
-    /// ```rust,no_run
-    /// use crate::rstrace::hdr_image::HDR;
-    /// use std::fs::File;
-    /// use std::io::BufWriter;
-    /// use std::net::TcpStream;
-    /// use endianness::ByteOrder;
-    ///
-    ///
-    /// # fn main() -> anyhow::Result<()> {
-    /// let img = HDR::new(1920, 1080);
-    /// let endianness = ByteOrder::LittleEndian;
-    ///
-    /// // Example 1: Writing to a physical file on disk (using BufWriter for optimal performance)
-    /// let file = File::create("render.pfm")?;
-    /// let mut disk_writer = BufWriter::new(file);
-    /// img.write_pfm(&mut disk_writer, &endianness)?;
-    ///
-    /// // Example 2: Writing directly to RAM (useful for automated tests or passing data to other APIs)
-    /// let mut memory_buffer: Vec<u8> = Vec::new();
-    /// img.write_pfm(&mut memory_buffer, &endianness)?;
-    ///
-    /// // Example 3: Streaming the image over a network connection (e.g., to a render node)
-    /// let mut network_stream = TcpStream::connect("127.0.0.1:8080")?;
-    /// img.write_pfm(&mut network_stream, &endianness)?;
-    /// # Ok(())
-    /// # }
-    /// ```
-    ///
-    /// # Notes
-    /// - The image is written in **binary format**
-    /// - Pixels are stored **bottom-to-top** as required by the PFM specification
-    /// - The scale factor encodes endianness:
-    ///   - Negative = little endian
-    ///   - Positive = big endian
-    pub fn write_pfm<W: Write>(&self, mut writer: W, endianness: &ByteOrder) -> Result<()> {
-        write!(
-            writer,
-            "PF\n{} {}\n{:.1}\n",
-            self.width,
-            self.height,
-            endianness_number(endianness)
-        )?;
-
-        match endianness {
-            ByteOrder::LittleEndian => {
-                for y in (0..self.height).rev() {
-                    for x in 0..self.width {
-                        let color = self.get_pixel(x, y)?;
-                        writer.write_f32::<LittleEndian>(color.r)?;
-                        writer.write_f32::<LittleEndian>(color.g)?;
-                        writer.write_f32::<LittleEndian>(color.b)?;
-                    }
-                }
-            }
-            ByteOrder::BigEndian => {
-                for y in (0..self.height).rev() {
-                    for x in 0..self.width {
-                        let color = self.get_pixel(x, y)?;
-                        writer.write_f32::<BigEndian>(color.r)?;
-                        writer.write_f32::<BigEndian>(color.g)?;
-                        writer.write_f32::<BigEndian>(color.b)?;
-                    }
-                }
-            }
-        }
-
-        Ok(())
-    }
-
     /// Sets the color of a pixel at `(x, y)`.
     ///
     /// # Errors
@@ -183,7 +99,6 @@ impl HDR {
     /// # Errors
     /// Returns an error if `(x, y)` is out of bounds.
     pub fn get_pixel(&self, x: usize, y: usize) -> Result<Color> {
-        //Ok(self.pixels[y * self.width + x])  Is it better this previous version??
         Ok(self.pixels[self.vector_index(x, y)?])
     }
 
@@ -209,9 +124,94 @@ impl HDR {
     }
 }
 
-// ====================
-//     Tone Mapping
-// ====================
+// =================================================================
+// PFM Writing
+// =================================================================
+
+impl HDR {
+    /// Writes the image to a `.pfm` (Portable Float Map) file.
+    ///
+    /// # Arguments
+    /// * `writer` - The output destination (e.g., a file, a memory buffer, or a network socket).
+    ///   It accepts any type that implements the [`Write`] trait.
+    /// * `endianness` - Byte order used for writing floats
+    ///
+    /// # Errors
+    /// Returns an error if:
+    /// - The underlying stream (`writer`) cannot be written to.
+    /// - An I/O error occurs while formatting the header or encoding the pixels.
+    ///
+    /// # Examples
+    ///
+    /// ```rust,no_run
+    /// use rstrace::hdr_image::HDR;
+    /// use std::fs::File;
+    /// use std::io::BufWriter;
+    /// use endianness::ByteOrder;
+    ///
+    ///
+    /// # fn main() -> anyhow::Result<()> {
+    /// use rstrace::pfm_func::Endianness;
+    /// let img = HDR::new(1920, 1080);
+    /// let endianness = Endianness::LittleEndian;
+    ///
+    /// // Example 1: Writing to a physical file on disk (using BufWriter for optimal performance)
+    /// let file = File::create("render.pfm")?;
+    /// let mut disk_writer = BufWriter::new(file);
+    /// img.write_pfm(&mut disk_writer, &endianness)?;
+    ///
+    /// // Example 2: Writing directly to RAM (useful for automated tests or passing data to other APIs)
+    /// let mut memory_buffer: Vec<u8> = Vec::new();
+    /// img.write_pfm(&mut memory_buffer, &endianness)?;
+    /// # Ok(())
+    /// # }
+    /// ```
+    ///
+    /// # Notes
+    /// - The image is written in **binary format**
+    /// - Pixels are stored **bottom-to-top** as required by the PFM specification
+    /// - The scale factor encodes endianness:
+    ///   - Negative = little endian
+    ///   - Positive = big endian
+    pub fn write_pfm<W: Write>(&self, mut writer: W, endianness: &Endianness) -> Result<()> {
+        write!(
+            writer,
+            "PF\n{} {}\n{:.1}\n",
+            self.width,
+            self.height,
+            endianness_number(endianness)
+        )?;
+
+        match endianness {
+            Endianness::LittleEndian => {
+                for y in (0..self.height).rev() {
+                    for x in 0..self.width {
+                        let color = self.get_pixel(x, y)?;
+                        writer.write_f32::<LittleEndian>(color.r)?;
+                        writer.write_f32::<LittleEndian>(color.g)?;
+                        writer.write_f32::<LittleEndian>(color.b)?;
+                    }
+                }
+            }
+            Endianness::BigEndian => {
+                for y in (0..self.height).rev() {
+                    for x in 0..self.width {
+                        let color = self.get_pixel(x, y)?;
+                        writer.write_f32::<BigEndian>(color.r)?;
+                        writer.write_f32::<BigEndian>(color.g)?;
+                        writer.write_f32::<BigEndian>(color.b)?;
+                    }
+                }
+            }
+        }
+
+        Ok(())
+    }
+}
+
+// =================================================================
+// Tone Mapping
+// =================================================================
 
 impl HDR {
     /// Computes the logarithmic average luminance of the image.
@@ -278,7 +278,7 @@ impl HDR {
         }
 
         let avr = self.average_luminosity()?;
-        if avr == 0.0 {
+        if are_close(avr, 0.0) {
             return Err(anyhow!(
                 "normalization():
             Average luminosity is zero, cannot normalize."
@@ -291,7 +291,7 @@ impl HDR {
         Ok(())
     }
 
-    /// Applies Tone mapping to all pixels.
+    /// Applies tone mapping to all pixels.
     ///
     /// # Errors
     /// Returns an error if the image is empty.
@@ -340,6 +340,7 @@ impl HDR {
 /// - Pixel values are assumed to be in row-major order
 /// - Gamma correction is applied as `x^(1/gamma)`
 /// - Output values are clamped to `[0, 1]` before conversion to `[0, 255]`
+/// - Negative values stored in HDR will produce `NaN`.
 ///
 /// # Example
 /// ```rust, no_run
@@ -397,13 +398,15 @@ pub fn hdr_to_ldr(argv: &mut Parameter) -> Result<()> {
 
     Ok(())
 }
-//                 tests
+
+// =================================================================
+// Tests
+// =================================================================
 
 #[cfg(test)]
 mod test {
     use super::*;
     use crate::functions::are_close;
-    // Test for
     #[test]
     fn test_new() {
         let hdr = HDR::new(10, 55);
@@ -418,7 +421,7 @@ mod test {
     }
 
     #[test]
-    fn test_set_pixel() {
+    fn test_get_and_set_pixel() {
         let mut hdr = HDR::new(10, 2);
         hdr.set_pixel(
             5,
@@ -435,23 +438,9 @@ mod test {
         assert_eq!(pixel.g, 2.5);
         assert_eq!(pixel.b, 10.0);
     }
-    #[test]
-    fn test_get_pixel() {
-        let mut hdr = HDR::new(10, 2);
-        let color = Color {
-            r: 1.0,
-            g: 0.2,
-            b: 30.0,
-        };
-        hdr.set_pixel(1, 1, color).unwrap();
-        let pixel = hdr.get_pixel(1, 1).unwrap();
-        assert_eq!(pixel.r, color.r);
-        assert_eq!(pixel.g, color.g);
-        assert_eq!(pixel.b, color.b);
-    }
 
     #[test]
-    #[should_panic]
+    #[should_panic(expected = "OUT OF BOUND PIXEL")]
     fn test_get_pixel_panic() {
         let hdr = HDR::new(10, 2);
         let _ = hdr.get_pixel(11, 1).unwrap();
@@ -466,7 +455,7 @@ mod test {
     }
 
     #[test]
-    #[should_panic]
+    #[should_panic(expected = "OUT OF BOUND PIXEL")]
     fn test_check_position() {
         let hdr = HDR::new(10, 55);
         hdr.check_position(11, 2).unwrap();
@@ -494,23 +483,23 @@ mod test {
         let mut img = HDR::new(3, 2);
 
         img.set_pixel(0, 0, Color::new(1.0e1, 2.0e1, 3.0e1))
-            .unwrap(); // Each component is
+            .unwrap();
         img.set_pixel(1, 0, Color::new(4.0e1, 5.0e1, 6.0e1))
-            .unwrap(); // different from any
+            .unwrap();
         img.set_pixel(2, 0, Color::new(7.0e1, 8.0e1, 9.0e1))
-            .unwrap(); // other: important in
+            .unwrap();
         img.set_pixel(0, 1, Color::new(1.0e2, 2.0e2, 3.0e2))
-            .unwrap(); // tests!
+            .unwrap();
         img.set_pixel(1, 1, Color::new(4.0e2, 5.0e2, 6.0e2))
             .unwrap();
         img.set_pixel(2, 1, Color::new(7.0e2, 8.0e2, 9.0e2))
             .unwrap();
         let mut buffer: Vec<u8> = vec![];
-        img.write_pfm(&mut buffer, &ByteOrder::LittleEndian)
+        img.write_pfm(&mut buffer, &Endianness::LittleEndian)
             .unwrap();
         assert_eq!(buffer, reference_le_bytes);
         buffer = vec![];
-        img.write_pfm(&mut buffer, &ByteOrder::BigEndian).unwrap();
+        img.write_pfm(&mut buffer, &Endianness::BigEndian).unwrap();
         assert_eq!(buffer, reference_be_bytes);
     }