+++ src/pfm_func.rs
@@ -26,7 +26,7 @@
 //! - Extra trailing bytes are treated as an error.
 //! - Pixel data is stored in row-major order.
 use crate::color::Color;
-use crate::hdr_image::HDR;
+use crate::hdr_image::{HDR, hdr_to_ldr};
 use anyhow::anyhow;
 use std::fs::File;
 use std::io::{BufRead, BufReader, Read};
@@ -323,7 +323,7 @@ impl Parameter {
     /// # Example
     /// ```rust, no_run
     /// use rstrace::pfm_func::Parameter;
-    /// let args = vec![
+    /// let args: Vec<String> = vec![
     ///     "program".into(),
     ///     "input.pfm".into(),
     ///     "0.18".into(),
@@ -331,21 +331,21 @@ impl Parameter {
     ///     "output.png".into(),
     /// ];
     ///
-    /// let params = Parameter::new(args).unwrap();
+    /// let params = Parameter::new(&args).unwrap();
     /// ```
-    pub fn new(args: Vec<String>) -> anyhow::Result<Parameter> {
-        if args.len() != 5 {
+    pub fn new(args: &Vec<String>) -> anyhow::Result<Parameter> {
+        if args.len() != 4 {
             return Err(anyhow!(
                 "wrong number of parameters: expected\n\
             <input_file_name> <factor_a> <gamma> <output_file_name>"
             ));
         }
 
-        let input_temp: &String = &args[1];
+        let input_temp: &String = &args[0];
         let input_pfm_file_name = input_temp.to_string();
-        let mut factor_a: f32 = args[2].parse::<f32>().expect("invalid factor_a value");
-        let mut gamma: f32 = args[3].parse::<f32>().expect("invalid gamma value");
-        let output_temp: &String = &args[4];
+        let mut factor_a: f32 = args[1].parse::<f32>().expect("invalid factor_a value");
+        let mut gamma: f32 = args[2].parse::<f32>().expect("invalid gamma value");
+        let output_temp: &String = &args[3];
         let output_file_name: String = output_temp.to_string();
         if factor_a <= 0.0 {
             println!("factor 'a' was automatically set to 0.18");
@@ -364,6 +364,30 @@ impl Parameter {
             output_file_name,
         })
     }
+}
+
+pub fn pfm_to_png(
+    input_file: String,
+    factor_a: f32,
+    gamma: f32,
+    output_file: String,
+) -> anyhow::Result<()> {
+    let args = vec![
+        input_file,
+        factor_a.to_string(),
+        gamma.to_string(),
+        output_file,
+    ];
+
+    let mut params = Parameter::new(&args)?;
+
+    let file = File::open(&args[0]);
+    let mut reader: BufReader<File> = BufReader::new(file?);
+    let mut img = read_pfm(&mut reader)?;
+    img.normalization(Some(&factor_a))?;
+    img.sem_clamp_image()?;
+    hdr_to_ldr(&mut params)?;
+    Ok(())
 }
 
 #[cfg(test)]
@@ -471,39 +495,39 @@ mod test {
     #[should_panic]
     //should panic if factor_a is not a number
     fn test_1_new_parameter() {
-        let strings: Vec<String> = ["exe", "filename_in", "a", "2.2", "filename_out"]
+        let strings: Vec<String> = ["filename_in", "a", "2.2", "filename_out"]
             .map(String::from)
             .to_vec();
-        let _par = Parameter::new(strings);
+        let _par = Parameter::new(&strings);
     }
 
     #[test]
     #[should_panic]
     //should panic if gamma is not a number
     fn test_2_new_parameter() {
-        let strings: Vec<String> = ["exe", "filename_in", "0.18", "a", "filename_out"]
+        let strings: Vec<String> = ["filename_in", "0.18", "a", "filename_out"]
             .map(String::from)
             .to_vec();
-        let _par = Parameter::new(strings);
+        let _par = Parameter::new(&strings);
     }
 
     #[test]
     //sets factor_a to 0.18 when a < 0
     fn test_3_new_parameter() {
-        let strings: Vec<String> = ["exe", "filename_in", "-1", "2.2", "filename_out"]
+        let strings: Vec<String> = ["filename_in", "-1", "2.2", "filename_out"]
             .map(String::from)
             .to_vec();
-        let par = Parameter::new(strings).unwrap();
+        let par = Parameter::new(&strings).unwrap();
         assert_eq!(0.18, par.factor_a);
     }
 
     #[test]
     //sets gamma to 2.2 when gamma < 0
     fn test_4_new_parameter() {
-        let strings: Vec<String> = ["exe", "filename_in", "0.18", "-1", "filename_out"]
+        let strings: Vec<String> = ["filename_in", "0.18", "-1", "filename_out"]
             .map(String::from)
             .to_vec();
-        let par = Parameter::new(strings).unwrap();
+        let par = Parameter::new(&strings).unwrap();
         assert_eq!(2.2, par.gamma);
     }
 
@@ -511,17 +535,10 @@ mod test {
     #[should_panic]
     //should panic if incorrect number of input parameters
     fn test_5_new_parameter() {
-        let strings: Vec<String> = [
-            "added string",
-            "exe",
-            "filename_in",
-            "0.18",
-            "a",
-            "filename_out",
-        ]
-        .map(String::from)
-        .to_vec();
-        let _par = Parameter::new(strings).unwrap();
+        let strings: Vec<String> = ["added string", "filename_in", "0.18", "a", "filename_out"]
+            .map(String::from)
+            .to_vec();
+        let _par = Parameter::new(&strings).unwrap();
     }
 
     #[test]